requirements.txt
Flask
azure-storage-blob
python-dotenv
groq-python-sdk